ClassUtil had a method called changeAnnotationValue to change the value of an annotation field value. This method made changes to an internal data structure that is not intended to be altered. From Java-13 onwards a warning appears in the console for manipulating this data structure. Therefore the method has been removed.
The CliUtil class in djutils-ext used the changedAnnotationValue method from ClassUtil. Instead of changing the internal data structure of the annotations, an override map is now used to store the changes. This also ensures that the code structure is reentrant: no structural changes have been made to the appearance of the loaded code (which the changeAnnotationValue method did in a sense).