The float and char primitive types were missing as allowed payloads for the fireEvent and fireTimedEvent methods. Methods have been added to the classes EventProducer, EventProducerImpl and RemoteEventProducer.
It is now possible to turn off verification of the fields of each individual event, which can be useful for
high-frequency events. The method to verify uses the isAssignableFrom(...) method that could be quite 
expensive. The fireEvent methods in EventProducerImpl have an extra boolean for verification (true by default).  

The EventProducer, EventProducerImpl and RemoteEventProducer got extra methods to be able to fire
unverified events as well.