EventListener and EventProducer interfaces are now Remote.
Since Java 8u241, a not so well documented change has been made in
Remote classes for RMI. See https://bugs.openjdk.org/browse/JDK-8237213:
Only those methods specified in a "remote interface", an interface that
**DIRECTLY** extends java.rmi.Remote are available remotely. This means
that when the Remote interface is added in a sub-interface, methods are
NOT available remotely. Repeating the method declaration in the
interface that extends Remote will correctly indicate the method is a
remote method. The base interface does not need to extend Remote if the
method is re-declared in the Remote interface.