EventProducer interface also implements all fireEvent() methods.
This is to make it easier for classes implementing the EventProducer
interface to immediately have all the publisher's behavior available.
All methods throw RemoteException; the LocalEventProducer (which only
needs to implement 1 real method now) wraps all methods with a
try-catch, as no RemoteExceptions are to be expected from a
LocalEventProducer. If such an Exception would happen, it wraps it in a
RuntimeException.