Issue #8. Make it easier to deserialize non-djunits classes
An extra interface for the Serializers has been added for types that do
not need to be specified further, nor need a unit. To avoid having to
implement IllegalOperationException, the djunits versions of the
TextSerializer are taken as the basis, and an extra interface with a
convenience deserialize(text) method is added. This interface is called
SpecificTextSerializer. All primitive serializers and the
StringSerializer implement the SpecificTextSerializer.